% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausmacrodata.R
\name{ausmacro}
\alias{ausmacro}
\title{Read data from ausmacrodata.org}
\usage{
ausmacro(dataset, format = c("ts", "tibble"), ...)
}
\arguments{
\item{dataset}{a URL for a specific data set on ausmacrodata.org, or an id for a data set.}

\item{format}{The desired format for the object to be returned.}

\item{...}{Other arguments, not currently used.}
}
\description{
\code{ausmacro} returns a data set from \url{ausmacrodata.org}.
By default, it will return a time series (a `ts` object), although
a `tibble` is also possible (and probably preferable for daily data).
}
\examples{
y <- ausmacro("http://ausmacrodata.org/series.php?id=gdpcknaasaq")
plot(y)
z <- ausmacro("gdpcknaasaq")
}
\author{
Rob J Hyndman
}
