% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_shape.R
\name{download_shape}
\alias{download_shape}
\title{Download shape file}
\usage{
download_shape(url)
}
\arguments{
\item{url}{path to zip file to be downloaded.}
}
\value{
a shape file
}
\description{
Function for downloading shape file. Downloads zip file, unzips and loads
.shp file.
}
