% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_shapefile.R
\name{download_shapefile}
\alias{download_shapefile}
\title{Download shape file}
\usage{
download_shapefile(url)
}
\arguments{
\item{url}{path to zip file to be downloaded.}

\item{output_type}{select from "sf" for simple format or "shape" for shape
file.}
}
\value{
returns either a sf or shape data frame object.
}
\description{
Function for downloading shape files. Downloads zip file, unzips and loads
.shp file.
}
\examples{
metadata <- ckanr::package_show("062801f2-dcf9-4a2d-b65b-52f20d4da721",
                                as = "table",
                                url = "http://www.data.gov.au")
url <- metadata$resources[["url"]]
dataset <- download_shapefile(url)
}
\author{
Cameron Roach
}
