% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ausmacrodata.R
\name{get_ausmacrodata}
\alias{get_ausmacrodata}
\title{Read data from ausmacrodata.org}
\usage{
get_ausmacrodata(dataset, format)
}
\arguments{
\item{dataset}{a URL for a specific data set on ausmacrodata.org, or an id for a data set.}

\item{format}{The desired format for the object to be returned. Possible
values are \code{ts}, \code{xts} and \code{tibble}. By default a \code{ts}
object is returned for quarterly, monthly or annual data, and an
\code{xts} object is returned for daily data.}
}
\description{
\code{get_ausmacrodata} returns a time series data set from \url{http://ausmacrodata.org}.
}
\examples{
# Read quarterly Australia GDP using full URL
y <- get_ausmacrodata("http://ausmacrodata.org/series.php?id=gdpcknaasaq")
plot(y)

# Read daily cash rate target interest rate using only series ID
z <- get_ausmacrodata('crtdoirymmdir')
library(xts)
plot(z)
}
\author{
Rob J Hyndman
}
