% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oz_dataset.R
\name{get_oz_dataset}
\alias{get_oz_dataset}
\title{Get data set from Australian data portal}
\usage{
get_oz_dataset(x, progress, ...)
}
\arguments{
\item{x}{\code{character} or \code{data.frame} object that contains
information on which data set to download. See Details for more
information.}

\item{progress}{\code{character} passed to \code{.progress} argument in
\code{\link[plyr]{llply}}.}

\item{...}{arguments passed to \code{\link[plyr]{llply}}.}
}
\value{
\code{list} of data.
}
\description{
Get data set from Australian data portal
}
\details{
If argument to \code{x} is a \code{character} then the
 input is assumed to be the data set id. If it is a \code{data.frame}
 then the ids are accessed from it \code{data.frame}.
}
\examples{
\donttest{
# download data using character id
d <- get_oz_dataset("062801f2-dcf9-4a2d-b65b-52f20d4da721")

# download data using data.frame
d2 <- get_oz_dataset(oz_metadata[1:2,])
}
}
