% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url.R
\name{get_url}
\alias{get_url}
\title{Get the download URL for an Oz dataset using its ID}
\usage{
get_url(id)
}
\arguments{
\item{id}{ID for which to find a URL}
}
\value{
The download URL for the item for which the ID has been given
}
\description{
Function for finding the download URL given a data set ID.
}
\examples{
get_url(id = "062801f2-dcf9-4a2d-b65b-52f20d4da721")

}
\author{
Adam H Sparks
}
