% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_dataset.R
\name{get_url_dataset}
\alias{get_url_dataset}
\title{Get a data set from a URL}
\usage{
get_url_dataset(x, force = FALSE)
}
\arguments{
\item{x}{\code{character} URL containing data to download and import.}

\item{force}{\code{logical} If the argument to \code{x} contains a data set
that is greater than 100 Mb in size, should the data set still be
downloaded?.}
}
\value{
\code{\link[tibble]{tibble}} or \code{\link[sf]{sfc}}
  object.
}
\description{
Download and import a data set from a URL. This function assumes that the
URL points to a zip archive.
}
\examples{
url <- paste0("https://datagovau.s3.amazonaws.com/bioregionalassessm",
              "ents/LEB/PED/DATA/Resources/Other/110_Context_statement",
              "NTGAB_Selected_Wetlands/062801f2-dcf9-4a2d-b65b-52f20d4",
              "da721.zip")
dataset <- get_url_dataset(url)

}
