% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oz_metadata}
\alias{oz_metadata}
\title{Meta data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2004 rows and 12 columns.}
\usage{
data(oz_metadata)
}
\description{
Meta-data for all data on \url{http://data.gov.au}.
}
\keyword{data}
