% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{ozdata}
\alias{ozdata}
\alias{ozdata-package}
\title{ozdata: Import and explore freely available Australian data sets}
\description{
Obtain data from data.gov.au and ozmacrodata.org, process it, plot it.
}
\details{
Created as part of the 2017 BURGr R UnConference in Brisbane, Australia.

Authors:
\itemize{
\item \href{https://github.com/jonocarroll}{Jonathan Carroll}
\item \href{https://github.com/marianess}{Maria Dmitrijeva}
\item \href{https://github.com/jeffreyhanson}{Jeffrey O. Hanson}
\item \href{https://github.com/robjhyndman}{Rob J. Hyndman}
\item \href{https://github.com/SimonLyons}{Simon Lyons}
\item \href{https://github.com/camroach87}{Cameron Roach}
\item \href{https://github.com/adamhsparks}{Adam H Sparks}
\item \href{https://github.com/matt-sutton}{Matt Sutton}
}

For a general overview of the steps taken to produce this, see \code{vignette("ozdata")}.
}
