% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_search.R
\name{run_shiny_search}
\alias{run_shiny_search}
\title{Run shiny search app}
\usage{
run_shiny_search()
}
\value{
A shiny application
}
\description{
This function runs the shiny search app for data.gov.au.
}
\examples{
\donttest{
run_shiny_search()
}
}
\author{
Simon Lyons, Cameron Roach
}
