% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_metadata.R
\name{search_metadata}
\alias{search_metadata}
\title{search_metadata}
\usage{
search_metadata(metadata = NULL, search)
}
\arguments{
\item{metadata}{a metadata table to search}

\item{search}{the search term to check}
}
\value{
a \code{tibble} of metadata rows which match the search
}
\description{
Runs \code{grepl} over the \code{tags} column of the metadata file and returns matching rows
}
