% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_oz_metadata.R
\name{download_oz_metadata}
\alias{download_oz_metadata}
\title{Download meta-data}
\usage{
download_oz_metadata(max_results = 1000)
}
\arguments{
\item{max_results}{Maximum results to return. Set to \code{Inf} for everything (40,000+).}
}
\value{
\code{link[tibble]{tibble}} table.
}
\description{
Meta-data for all Australian datasets on \url{https://data.gov.au}. Collecting all of the meta-data entries will take quite some time.
}
\examples{
\donttest{
# download latest version of metadata
metadata <- download_oz_metadata(max_results = 100)
# print first 10 rows of data set
head(metadata)
}
}
