% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_shape.R
\name{download_shape}
\alias{download_shape}
\title{Download shape file}
\usage{
download_shape(url, output_type = "sf")
}
\arguments{
\item{url}{path to zip file to be downloaded.}

\item{output_type}{select from "sf" for simple format or "shape" for shape
file.}
}
\value{
returns either a sf or shape data frame object.
}
\description{
Function for downloading shape files. Downloads zip file, unzips and loads
.shp file.
}
\examples{
ckanr_setup("http://www.data.gov.au")
url <- package_show('062801f2-dcf9-4a2d-b65b-52f20d4da721', as = 'table')$resources[["url"]]
shape <- download_shape(url)

}
\author{
Cameron Roach
}
